+++ R/pltree.R
@@ -100,16 +100,12 @@ pltree <- function (formula, data, subset, na.action, cluster, ref = NULL, ...){
     m <- m[!control_args]
     m$control <- control
     m$fit <- as.name("plfit")
-    m[[1L]] <- as.name("mob")
+    m[[1L]] <- quote(partykit::mob)
     rval <- eval(m, parent.frame())
     rval$info$call <- m
     class(rval) <- c("pltree", "bttree", class(rval))
     return(rval)
 }
-
-#' @importFrom partykit mob
-#' @export
-partykit::mob
 
 #' @method print pltree
 #' @export
@@ -121,11 +117,11 @@ print.pltree <- function (x, title = "Plackett-Luce tree",
 }
 
 #' @method plot pltree
+#' @importFrom psychotree node_btplot
 #' @export
-plot.pltree <- function (x, terminal_panel = psychotree::node_btplot,
+plot.pltree <- function (x, terminal_panel = node_btplot,
                          tp_args = list(...),
                          tnex = NULL, drop_terminal = NULL, ...){
-    requireNamespace("psychotree")
     NextMethod()
 }
 
@@ -179,6 +175,7 @@ coef.pltree <- function (object, node = NULL, drop = TRUE, ...) {
 #' @method itempar pltree
 #' @export
 itempar.pltree <- function (object, ...){
+    # so unexported itempar.bttree is used from psychotree
     requireNamespace("psychotree")
     NextMethod()
 }