+++ NEWS.md
@@ -1,3 +1,5 @@
+# PlackettLuce Development version
+
 # PlackettLuce 0.2-2
 
 ## New features