+++ NEWS.md
@@ -1,5 +1,10 @@
 # PlackettLuce Development version
 
+## Improvements
+
+* Print methods for `"PlackettLuce"` and `"summary.PlacketLuce"` objects now respect `options("width")`.
+
+
 ## Bug fixes
 
 * Correct vcov for weighted rankings of more than two items.
@@ -9,8 +14,7 @@
 
 ## New features
 
-* Add `AIC.pltree` to enable computation of AIC on new observations (e.g.
-data held out in cross-validation).
+* Add `AIC.pltree` to enable computation of AIC on new observations (e.g. data held out in cross-validation).
 * Add `fitted.pltree` to return combined fitted probabilities for each choice within each ranking, for each node in a Plackett-Luce tree.
 
 ## Bug fixes